feat: Add dochia list --profiles to display the built in profiles

+++ src/main/java/dev/dochia/cli/core/model/TestReport.java
@@ -17,10 +17,10 @@ import lombok.Getter;
 @Builder
 public class TestReport {
     private final List<TestCaseSummary> testCases;
-    private final int totalTests;
-    private final int success;
-    private final int warnings;
-    private final int errors;
+    private final long totalTests;
+    private final long success;
+    private final long warnings;
+    private final long errors;
     private final long executionTime;
     private final String timestamp;
     private final String dochiaVersion;
