+++ src/main/java/dev/dochia/cli/core/report/ExecutionStatisticsListener.java
@@ -19,17 +19,17 @@ public class ExecutionStatisticsListener {
     /**
      * Map to track the count of errors per path.
      */
-    private final Map<String, Integer> errors = new HashMap<>();
+    private final Map<String, Long> errors = new HashMap<>();
 
     /**
      * Map to track the count of warnings per path.
      */
-    private final Map<String, Integer> warns = new HashMap<>();
+    private final Map<String, Long> warns = new HashMap<>();
 
     /**
      * Map to track the count of successful executions per path.
      */
-    private final Map<String, Integer> success = new HashMap<>();
+    private final Map<String, Long> success = new HashMap<>();
 
     /**
      * Count of skipped tests.
@@ -76,7 +76,7 @@ public class ExecutionStatisticsListener {
      * @param path The path for which errors are increased.
      */
     public void increaseErrors(String path) {
-        this.errors.merge(path, 1, Integer::sum);
+        this.errors.merge(path, 1L, Long::sum);
     }
 
     /**
@@ -85,7 +85,7 @@ public class ExecutionStatisticsListener {
      * @param path The path for which warnings are increased.
      */
     public void increaseWarns(String path) {
-        this.warns.merge(path, 1, Integer::sum);
+        this.warns.merge(path, 1L, Long::sum);
     }
 
     /**
@@ -94,7 +94,7 @@ public class ExecutionStatisticsListener {
      * @param path The path for which successful executions are increased.
      */
     public void increaseSuccess(String path) {
-        this.success.merge(path, 1, Integer::sum);
+        this.success.merge(path, 1L, Long::sum);
     }
 
     /**
@@ -102,8 +102,8 @@ public class ExecutionStatisticsListener {
      *
      * @return The total count of errors.
      */
-    public int getErrors() {
-        return this.errors.values().stream().reduce(0, Integer::sum);
+    public long getErrors() {
+        return this.errors.values().stream().reduce(0L, Long::sum);
     }
 
     /**
@@ -111,8 +111,8 @@ public class ExecutionStatisticsListener {
      *
      * @return The total count of warnings.
      */
-    public int getWarns() {
-        return this.warns.values().stream().reduce(0, Integer::sum);
+    public long getWarns() {
+        return this.warns.values().stream().reduce(0L, Long::sum);
     }
 
     /**
@@ -120,8 +120,8 @@ public class ExecutionStatisticsListener {
      *
      * @return The total count of successful executions.
      */
-    public int getSuccess() {
-        return this.success.values().stream().reduce(0, Integer::sum);
+    public long getSuccess() {
+        return this.success.values().stream().reduce(0L, Long::sum);
     }
 
     /**
@@ -129,7 +129,7 @@ public class ExecutionStatisticsListener {
      *
      * @return The total count of all executions.
      */
-    public int getAll() {
+    public long getAll() {
         return this.getSuccess() + this.getWarns() + this.getErrors();
     }
 
@@ -158,9 +158,9 @@ public class ExecutionStatisticsListener {
      * @return A formatted string representation of the execution results.
      */
     public String resultAsStringPerPath(String path) {
-        String errorsString = AnsiUtils.boldRed("E " + errors.getOrDefault(path, 0));
-        String warnsString = AnsiUtils.boldYellow("W " + warns.getOrDefault(path, 0));
-        String successString = AnsiUtils.boldGreen("S " + success.getOrDefault(path, 0));
+        String errorsString = AnsiUtils.boldRed("E " + errors.getOrDefault(path, 0L));
+        String warnsString = AnsiUtils.boldYellow("W " + warns.getOrDefault(path, 0L));
+        String successString = AnsiUtils.boldGreen("S " + success.getOrDefault(path, 0L));
         return "%s, %s, %s".formatted(errorsString, warnsString, successString);
     }
 
@@ -171,6 +171,6 @@ public class ExecutionStatisticsListener {
      * @return The total count of executions for the specified path.
      */
     public long getExecutionsPerPath(String path) {
-        return this.errors.getOrDefault(path, 0) + this.warns.getOrDefault(path, 0) + this.success.getOrDefault(path, 0);
+        return this.errors.getOrDefault(path, 0L) + this.warns.getOrDefault(path, 0L) + this.success.getOrDefault(path, 0L);
     }
 }
